add servlets and fix bugs

+++ src/main/java/UserPackage/UserServlet.java
@@ -4,62 +4,77 @@ package UserPackage;
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.*;
 import java.io.IOException;
 import java.sql.Date;
 import java.sql.SQLException;
 
-import FriendsPackage.FriendsDao;
-import FriendsPackage.FriendsSqlDao;
+import ProfilePackage.CookieManager;
 import ServletContextPackage.ContextDataNames;
 
 @WebServlet(name = "UserServlet", urlPatterns = {"/UserServlet"})
 public class UserServlet extends HttpServlet {
+    private static final String pageAddress = "/WEB-INF/views/";
 
     @Override
     protected void doPost(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) throws ServletException, IOException {
         ServletContext servletContext = getServletContext();
+        HttpSession session = httpServletRequest.getSession();
+        Cookie[] cookies = httpServletRequest.getCookies();
         UserSqlDao uDao = (UserSqlDao) servletContext.getAttribute(ContextDataNames.USER_DAO);
-        String name = httpServletRequest.getParameter("username");
-        int id = 0;
-        try {
-            id = uDao.getUserIdByName(name);
-        } catch (SQLException e) {
-            e.printStackTrace();
+        String userName = (String) session.getAttribute("currentUser");
+        if (userName == null) {
+            httpServletResponse.sendRedirect("/");
+            return;
         }
-        String webPageName = "";
         if (httpServletRequest.getParameter("button").equals("delete")) {
+            int id = 0;
+            try {
+                id = uDao.getUserIdByName(userName);
+            } catch (SQLException e) {
+                e.printStackTrace();
+            }
             try {
                 uDao.deleteUser(uDao.getUser(id));
-                webPageName = "DeletePage.jsp";
+                session.invalidate();
+                CookieManager.deleteMyCookies(httpServletResponse, cookies);
+                httpServletResponse.sendRedirect("/");
             } catch (SQLException | ClassNotFoundException e) {
                 e.printStackTrace();
             }
         } else if (httpServletRequest.getParameter("button").equals("edit profile")) {
-            webPageName = "EditProfilePage.jsp";
-        } else if (httpServletRequest.getParameter("button").equals("edit")) {
-            webPageName = "UserPage.jsp";
-            try {
-                uDao.addProfile(id, httpServletRequest.getParameter("name"), httpServletRequest.getParameter("surname"), httpServletRequest.getParameter("birthPlace"), httpServletRequest.getParameter("status"));
-            } catch (SQLException e) {
-                e.printStackTrace();
-            }
+            httpServletResponse.sendRedirect("/EditProfile");
         } else {
             try {
-                if (uDao.containsUserName(name)) {
-                    webPageName = "ProfilePage.jsp?id=" + id;
+                String searchName = httpServletRequest.getParameter("username");
+                int id = 0;
+                try {
+                    id = uDao.getUserIdByName(searchName);
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+                if (uDao.containsUserName(searchName)) {
+                    httpServletRequest.setAttribute("id", id);
+                    httpServletResponse.sendRedirect("/ProfilePage");
                 } else {
-                    webPageName = "UserNotFoundPage.jsp";
+                    httpServletRequest.setAttribute("error", "Username does not exist!");
                 }
             } catch (SQLException e) {
                 e.printStackTrace();
             }
         }
-
-
-        httpServletRequest.getRequestDispatcher(webPageName).forward(httpServletRequest, httpServletResponse);
     }
-    
+
+    @Override
+    protected void doGet(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) throws ServletException, IOException {
+        HttpSession session = httpServletRequest.getSession();
+        String currentUser = (String) session.getAttribute("currentUser");
+        String webPage = "";
+        if (currentUser == null) {
+            httpServletResponse.sendRedirect("/");
+            return;
+        }
+        webPage = pageAddress + "UserPage.jsp";
+        httpServletRequest.getRequestDispatcher(webPage).forward(httpServletRequest, httpServletResponse);
+    }
 }
